requirements.txt
# Core dependencies
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
django-extensions==3.2.3

# Channels for websockets and async support
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# Database
psycopg2-binary==2.9.9
django-timescaledb==0.2.13

# MQTT
paho-mqtt==2.1.0

# API
graphene-django==3.1.5
djangorestframework-simplejwt==5.3.1

# Frontend
django-htmx==1.17.2

# Development tools
pytest==7.4.3
pytest-django==4.7.0
black==23.12.1
isort==5.13.2
flake8==6.1.0

# Monitoring
prometheus-client==0.18.0

# Utilities
python-dotenv==1.0.1
faker==22.4.0
numpy==1.26.3
pandas==2.1.4

# API docs support
coreapi==2.3.3
coreschema==0.0.4
setuptools<81